replacing int for float

+++ Aula 4/desafio005.py
@@ -1,2 +1,2 @@
-n = int(input("Digite um número: "))
+n = float(input("Digite um número: "))
 print(f"O antecessor ao número digitado é: {n - 1}.\nE seu sucessor é o número: {n + 1}")
